- Add Base classes to tendrl namespace
tendrl-bug-id: Tendrl/specifications#37
tendrl-bug-id: Tendrl/specifications#54
tendrl-bug-id: Tendrl/specifications#32

jira: [TEN-3] [TEN-4]
