+++ tendrl/commons/jobs/__init__.py
@@ -89,10 +89,9 @@ class JobConsumer(object):
         if "tendrl" in flow_fqn and "objects" in flow_fqn:
             obj_name, flow_name = flow_fqn.split(".objects.")[-1].split(
                 ".flows.")
-            obj = tendrl_ns.get_object(obj_name)
             flow = tendrl_ns.get_obj_flow(obj_name, flow_name)
             return flow(parameters=job['parameters'],
-                        request_id=job['request_id'], obj=obj).run()
+                        request_id=job['request_id']).run()
 
         # flow_fqn eg: tendrl.node_agent.flows.temp_flows
         if "tendrl" in flow_fqn and "flows" in flow_fqn:
