+++ tendrl/commons/objects/__init__.py
@@ -1,0 +1,38 @@
+import abc
+import six
+
+
+@six.add_metaclass(abc.ABCMeta)
+class BaseObject(object):
+    def __init__(
+            self,
+            attrs=None,
+            enabled=None,
+            obj_list=None,
+            obj_value=None,
+            atoms=None,
+            flows=None
+    ):
+        self.attrs = attrs
+        self.enabled = enabled
+
+        # path to LIST of all instance of the object
+        self.obj_list = obj_list
+
+        # path to GET an instance of the object
+        self.obj_value = obj_value
+
+        self.atoms = atoms
+
+        # List of flows under this object
+        self.flows = flows
+
+    def __new__(cls, *args, **kwargs):
+
+        super_new = super(BaseObject, cls).__new__
+        if super_new is object.__new__:
+            instance = super_new(cls)
+        else:
+            instance = super_new(cls, *args, **kwargs)
+
+        return instance
