+++ tendrl/commons/persistence/etcd_persister.py
@@ -2,7 +2,5 @@ from tendrl.commons.persistence.persister import Persister
 
 
 class EtcdPersister(Persister):
-    def __init__(self, etcd_orm):
+    def __init__(self):
         super(EtcdPersister, self).__init__()
-        # Child classes responsible to provide valid etcdobj Server (orm)
-        self.etcd_orm = etcd_orm
