+++ tendrl/commons/sds_sync/__init__.py
@@ -1,6 +1,7 @@
 import abc
 
-import gevent
+import gevent.event
+import gevent.greenlet
 import six
